requirements.txt
streamlit==1.38.0
tensorflow==2.17.0
tensorflow-hub==0.16.1
Pillow==10.4.0
numpy==1.26.4